Add top level routing file and tests
Stub out the first iteration of method.js, which defines the entry
point for this module: handleRequest. This function will handle
checking the validity of the path and the resource itself, as well
as formatting of the parameters in the data object, and will call
into the appropriate action module.

Add failing functional tests for the various validity checks. Rework
the GET functional tests to go through the entry point in method.js
to be more end-to-end.